requirements.txt
setuptools~=68.0.0
requests~=2.31.0
click~=8.1.4
keyboard~=0.13.5
PyYAML~=6.0